A wrapper script for developer debugging of processonelog.sh.
Suggested use: download a directory's worth of errorlogs and run this script with the default (checked-in) processonelog.sh.  Then copy the resulting regresslogs.out to regresslogs.out.wanted and edit the results to taste.  (file format: "<errorlogname>: <reason>"). Then edit a local copy of processonelog.sh, switch this script to use it, and run it again.  You will be able to check your results in diff.out.